refactor: 주문 도메인을 분리하라
