+++ src/main/java/kitchenpos/order/table/infra/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.order.table.infra;
 
+import kitchenpos.order.table.domain.OrderTable;
+import kitchenpos.order.table.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
